requirements.txt
antlr4-python3-runtime==4.8
attrs==23.1.0
azure-core==1.29.5
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.2
diffprivlib==0.6.4
exceptiongroup==1.2.0
fonttools==4.43.1
gensim==4.3.2
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
greenlet==3.0.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-resources==6.1.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
msrest==0.7.1
networkx==3.1
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
opendp==0.8.0
opendp-smartnoise==0.1.4.2
opendp-smartnoise-core==0.2.6
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
pandasql==0.7.3
patsy==0.5.3
Pillow==10.1.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
portpicker==1.6.0
protobuf==3.20.3
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.16.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.16.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
statsmodels==0.14.0
threadpoolctl==3.3.0
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vl-convert-python==1.2.0
wsproto==1.2.0
zipp==3.17.0
